re-allocate sdp as a library
