+++ src/core/sdpsos/__init__.py
@@ -1,10 +1,10 @@
 from .sdpsos import SDPSOS, SOSProblem
 from .manifold import RootSubspace
-from .solver import (
+from .solution import SolutionSDP
+from ...sdp import (
     SDPProblem,
     SDPTransformation, SDPMatrixTransform, SDPRowMasking, SDPVectorTransform
 )
-from .solution import SolutionSDP
 
 __all__ = [
     'SDPSOS', 'SDPProblem', 'SOSProblem', 'RootSubspace',
