+++ src/core/structsos/quarternary/utils.py
@@ -9,7 +9,8 @@ from ...symsos import prove_univariate
 from ....utils.roots.rationalize import (
     nroots, rationalize, rationalize_bound, univariate_intervals
 )
-from ....utils.expression import congruence, CyclicExpr, CyclicSum, CyclicProduct, SymmetricSum, SymmetricProduct
+from ....sdp import congruence
+from ....utils import CyclicExpr, CyclicSum, CyclicProduct, SymmetricSum, SymmetricProduct
 
 a, b, c, d = sp.symbols("a b c d")
