+++ src/sdp/__init__.py
@@ -1,0 +1,23 @@
+# sdp.__init__.py
+#
+# This is a symbolic SDP library intended for independent use.
+
+from .arithmetic import (
+    solve_column_separated_linear, solve_undetermined_linear,
+    matmul, matmul_multiple, symmetric_bilinear, symmetric_bilinear_multiple
+)
+
+from .solver import (
+    SDPProblem,
+    SDPTransformation, SDPMatrixTransform, SDPRowMasking, SDPVectorTransform
+)
+
+from .utils import congruence
+
+__all__ = [
+    'solve_column_separated_linear', 'solve_undetermined_linear',
+    'matmul', 'matmul_multiple', 'symmetric_bilinear', 'symmetric_bilinear_multiple',
+    'SDPProblem',
+    'SDPTransformation', 'SDPMatrixTransform', 'SDPRowMasking', 'SDPVectorTransform',
+    'congruence'
+]
