+++ src/utils/expression/__init__.py
@@ -9,15 +9,11 @@ from .form import (
     poly_get_factor_form, poly_get_standard_form, latex_coeffs
 )
 
-from .solution import (
-    Solution, SolutionSimple,
-    congruence
-)
+from .solution import Solution, SolutionSimple
 
 __all__ = [
     'Coeff', 'identify_symmetry',
     'CyclicExpr', 'CyclicSum', 'CyclicProduct', 'SymmetricSum', 'SymmetricProduct', 'is_cyclic_expr',
     'poly_get_factor_form', 'poly_get_standard_form', 'latex_coeffs',
     'Solution', 'SolutionSimple',
-    'congruence',
 ]
